update id Integer to Long

+++ aidr-db-manager/src/main/java/qa/qcri/aidr/predictdb/entities/misc/Crisis.java
@@ -24,7 +24,7 @@ import javax.persistence.UniqueConstraint;
 @Table(name = "crisis", catalog = "aidr_predict", uniqueConstraints = @UniqueConstraint(columnNames = "code"))
 public class Crisis implements java.io.Serializable {
 
-	private Integer crisisId;
+	private Long crisisId;
 	private Users users;
 	private CrisisType crisisType;
 	private String name;
